feat/LOC5: Validating CRUD for MarcaController

+++ app/Http/Controllers/MarcaController.php
@@ -17,7 +17,7 @@ class MarcaController extends Controller
     public function index()
     {
         $marcas = $this->marca->all();
-        return $marcas;
+        return response()->json($marcas);
     }
     /**
      * Store a newly created resource in storage.
@@ -25,7 +25,7 @@ class MarcaController extends Controller
     public function store(Request $request)
     {
         $marca = $this->marca->create($request->all());
-        return $marca;
+        return response()->json($marca, 201);
     }
 
     /**
@@ -34,7 +34,10 @@ class MarcaController extends Controller
     public function show($id)
     {
         $marca = $this->marca->find($id);
-        return $marca;
+        if($marca === null){
+            return response()->json(['erro' => 'Nenhum registro encontrado'], 404);
+        }
+        return response()->json($marca);
     }
 
     /**
@@ -43,8 +46,11 @@ class MarcaController extends Controller
     public function update(Request $request, $id)
     {
         $marca = $this->marca->find($id);
+        if($marca === null){
+            return response()->json(['erro' => 'Nenhum registro encontrado'], 404);
+        }
         $marca->update($request->all());
-        return $marca;
+        return response()->json($marca);
     }
 
     /**
@@ -53,7 +59,10 @@ class MarcaController extends Controller
     public function destroy($id)
     {
         $marca = $this->marca->find($id);
+        if($marca === null){
+            return response()->json(['erro' => 'Nenhum registro encontrado'], 404);
+        }
         $marca->delete();
-        return $marca;
+        return response()->json($marca);
     }
 }
